assignment-3/requirements.txt
h5py==2.10.0
torch==1.5.0
matplotlib==3.2.1
pillow==7.1.2
cmake==3.17.2
dlib==19.19.0
